⬇ Moved views from init.py to different modules
